Add storelocator's test data to the app itself.
Note that Spanner doesn't seem to support the `schema.sql` and
`data.sql` approach, so an ApplicationRunner bean is used instead.

The initializeDatabase method was moved from the StoreService.

Also, there is a bug in spring-cloud-gcp-data-spanner, such that the
database creation incorrectly chooses the INT64 type for primitive
doubles. Hence the Store's latitude and longitude fields were updated to
Doubles.